Changed path for API calls in index.html of server to use relative path
The index.html (templates folder) of the server always calls webroot via `/api/...` this path will break (as it did for me) when you use to proxy to point at the Larynx server, e.g. `http://raspberrypi.local/sepia/larynx/`. I've changed the path to `api/...` and tested it for my Docker container. 
Note: this assumes that index.html is always served at the root path.